+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/PlayerHeaderFragment.kt
@@ -961,6 +961,7 @@ class PlayerHeaderFragment :
                 ArtworkOrVideo(
                     state = artworkOrVideoState,
                     onChapterUrlClick = {},
+                    videoSurfaceState = playbackManager.videoSurfaceState,
                     configureVideoView = { videoView ->
                         videoView.setOnClickListener { onFullScreenVideoClick() }
                     },
@@ -1110,6 +1111,7 @@ class PlayerHeaderFragment :
                         state = artworkOrVideoState,
                         artworkCornerRadius = if (availableHeightDp > 120.dp) 16.dp else 8.dp,
                         onChapterUrlClick = viewModel::onChapterUrlClick,
+                        videoSurfaceState = playbackManager.videoSurfaceState,
                         configureVideoView = { videoView ->
                             videoView.setOnClickListener { onFullScreenVideoClick() }
                         },
@@ -1167,6 +1169,7 @@ class PlayerHeaderFragment :
                 ) {
                     VideoBox(
                         player = artworkOrVideoState.player,
+                        videoSurfaceState = playbackManager.videoSurfaceState,
                         configureVideoView = { videoView ->
                             videoView.setOnClickListener { onFullScreenVideoClick() }
                         },